Search sources to TypeScript
